+++ src/Konmaripo.Web/Views/Home/Index.cshtml
@@ -1,83 +1,9 @@
-﻿@using Humanizer
-@model List<GitHubRepo>
-@{
-    var alreadyArchivedRepos = Model.Where(x => x.IsArchived).ToList();
-    var nonArchivedRepos = Model.Where(x => !x.IsArchived).OrderBy(x=>x.UpdatedDate).ToList();
+﻿@{
 
     ViewData["Title"] = "Home Page";
-
-    void RepoTable(List<GitHubRepo> repoList)
-    {
-        <table class="table table-striped">
-            <thead class="thead-dark">
-                <tr>
-                    <th scope="col"><!--Private icon--></th>
-                    <th scope="col">Name</th>
-                    <th scope="col" title="Stars"><i class="fas fa-star"></i></th>
-                    <th scope="col" title="Forks"><i class="fas fa-code-branch"></i></th>
-                    <th scope="col" title="Issues"><i class="fas fa-exclamation-circle"></i></th>
-                    <th scope="col" title="Date the repository was created">Created</th>
-                    <th scope="col" title="Date the repository was updated, e.g. title, language, settings, properties">Updated</th>
-                    <th scope="col" title="Date the Repository last had changes pushed to it">Pushed</th>
-                </tr>
-            </thead>
-            <tbody>
-             @foreach (var item in repoList)
-                {
-                <tr>
-                    @if (item.IsPrivate)
-                    {
-                        <td title="Private repository">
-                            <i class="fas fa-lock text-success"></i>
-                        </td>
-                    }
-                    else
-                    {
-                        <td title="Public repository">
-                            <i class="fas fa-lock-open text-danger"></i>
-                        </td>
-                    }
-                    <th scope="row" title="@item.Description"><a href="@item.RepoUrl" target="_blank"><i class="fas fa-external-link-alt"></i></a>&nbsp;@Html.ActionLink(item.Name, "RepositoryInfo", new {repoId = item.Id})</th>
-                    <td>@item.StarCount</td>
-                    <td>@item.ForkCount</td>
-                    <td>@item.OpenIssueCount</td>
-                    <td title="@item.CreatedDate">@item.CreatedDate.Humanize()</td>
-                    <td title="@item.UpdatedDate">@item.UpdatedDate.Humanize()</td>
-                    @if (item.PushedDate.HasValue)
-                    {
-                        <td title="@item.PushedDate">@item.PushedDate.Value.Humanize()</td>
-                    }
-                    else
-                    {
-                    <td>Never</td>
-                    }
-                </tr>
-                }
-            </tbody>
-        </table>
-    }
-
 }
 
 <div class="text-center">
-    <h1 class="display-4">Repository Overview</h1>
-    <p>Check out the repositories below to see if they could use some sunsetting.</p>
+    <h1 class="display-4">Welcome to Konmaripo!</h1>
+    <p>A tool to enable easier GitHub repo and organization management.</p>
 </div>
-
-<div class="text-center">
-    <h2>Active</h2>
-</div>
-
-@{ RepoTable(nonArchivedRepos); }
-
-
-
-<div class="text-center">
-    <h2>Archived</h2>
-</div>
-@{
-
-
-    RepoTable(alreadyArchivedRepos);
-}
-
